+++ private-endpoint/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Release v1.4.0
+
+## Provider & Terraform Upgrade
+- Azurerm provider: 3.96.0 (#13)
+- Terraform: 1.7.5 (#13)
+
+
+
+
+   
 ## Release v1.3.0
 
 ### Version upgrade